pyproject.toml
[tool.poetry]
name = "project"
version = "0.1.4"
description = ""
authors = ["Kodai YAMASHITA <nglcobdai@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.2"
uvicorn = "^0.29.0"
pydantic = "<2.0"
python-dotenv = "^1.0.1"
pymongo = "^4.7.2"
python-multipart = "^0.0.9"
openai-whisper = { git = "https://github.com/openai/whisper.git", rev = "v20231117" }
levenshtein = "^0.25.1"
pytz = "^2024.1"

[tool.poetry.group.dev.dependencies]
pep8 = "^1.7.1"
black = "^23.9.1"
flake8 = "^6.1.0"
isort = "^5.12.0"
ipython = "^8.15.0"
ipykernel = "^6.25.2"
pytest = "^8.2.0"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.flake8]
max-line-length = 120
